Decompose ChallengeDetail into participants, leaderboard, and log
Was 402 lines doing six things at once. Now:
  - components/ParticipantsTable.jsx (79) -- aggregated participants
    table with click-through to UserDetail
  - components/ChallengeLeaderboard.jsx (93) -- top-scorers card with
    medal chips and progress-vs-max bars; hidden when empty
  - components/ParticipationLog.jsx (93) -- per-completion event log
    with CSV export
  - ChallengeDetail.jsx (268) -- data loading, lookup helpers, header
    summary card, actions catalog, mobile-preview dialog

The ChallengeSummaryCard and aggregateParticipants helper stay inline
because they have no other consumer; pulling them out would only add
indirection. Behavior unchanged. Refs #42.
